uint64_t file_id/tStamp

+++ bitcask/Utils/utils.h
@@ -50,8 +50,8 @@ void scanHintFiles(std::vector<std::string>*);
 
 std::vector<std::string>* listDataFiles();
 
-uint32_t getLastFileInfo(std::vector<std::string>*);
+uint64_t getLastFileInfo(std::vector<std::string>*);
 
 int lockFile(std::string path);
 
-void writePID(int fd, uint32_t file_id);
+void writePID(int fd, uint64_t file_id);
